import the healthbar from HealthBars.js
